Restructuring of directories
