+++ mapServer/server/server_conf.py
@@ -1,0 +1,12 @@
+#!/usr/bin/python
+settings = {
+
+    'FILE_CONFIG':
+        {'filename':
+             r'/Users/empire/Academics/UCSC/nasaResearch/californiaNed30m/elevation_NED30M_ca_2925289_01/' \
+             r'virtRasterCalifornia.vrt'},
+
+    'HOST': 'localhost',
+    'PORT': 2002
+
+}
